server/requirements.txt
fastapi
python-dotenv
google-generativeai
uvicorn